bugfix about InterfaceWindow and VariableWindow
git-svn-id: https://ecell.svn.sourceforge.net/svnroot/ecell/ecell3/trunk@1535 f1531174-cb10-0410-9fe6-89aa7ac3eedb

+++ osogo/BoardWindow.py
@@ -127,12 +127,16 @@ class BoardWindow(OsogoWindow):
 			self.updatePositions()
 
 	def deletePluginWindowByTitle( self, aTitle ):
-		for aPluginFrame in self['board_table'].get_children():
-			if self.thePluginMap[str(aPluginFrame)].getTitle() == aTitle:
-				self.theSelectedPluginFrame = aPluginFrame
-				self.deletePluginWindow()
-				self.updatePositions()
-				break
+		if self.exists() == TRUE:
+			for aPluginFrame in self['board_table'].get_children():
+				if self.thePluginMap[str(aPluginFrame)].getTitle() == aTitle:
+					self.theSelectedPluginFrame = aPluginFrame
+					self.deletePluginWindow()
+					self.updatePositions()
+					break
+		# When this window is not created, does nothing.
+		else:
+			pass
 
 	def __radioButtonToggled( self, *arg ):
